Ready for demo - generator template (including price changes, av price change, standard deviation etc) complete. Next phase will be to use this to generate data. Possibly a need for SD of volumes/counts - to be discussed with advisor.

+++ src/generatorTemplate.cpp
@@ -9,7 +9,7 @@
 #include "iostream.h"
 using namespace std;
 
-generatorTemplate::generatorTemplate(string _wIssueSymbol, vector<double> _prices, vector<double> _priceChanges, double _avChange, double _sdChange, int _percentPositive, std::vector<int> _wTradeVolume, std::vector<int> _wTradeCount) {
+generatorTemplate::generatorTemplate(string _wIssueSymbol, vector<double> _prices, vector<double> _priceChanges, double _avChange, double _sdChange, double _percentPositive, std::vector<int> _wTradeVolume, std::vector<int> _wTradeCount) {
 	wIssueSymbol = _wIssueSymbol;
 	prices = _prices;
 	priceChanges = _priceChanges;
