+++ src/generatorTemplate.h
@@ -18,12 +18,12 @@ class generatorTemplate {
 	public: std::vector<double> priceChanges;
 	public: double avChange;
 	public: double sdChange;
-	public: int percentPositive;
+	public: double percentPositive;
 	public: std::vector<int> wTradeVolume;
 	public: std::vector<int> wTradeCount;
 	public: void print();
 public:
-	generatorTemplate(string, std::vector<double>, std::vector<double>, double, double, int, std::vector<int>, std::vector<int>);
+	generatorTemplate(string, std::vector<double>, std::vector<double>, double, double, double, std::vector<int>, std::vector<int>);
 	generatorTemplate();
 	generatorTemplate(const generatorTemplate&);
 	generatorTemplate& operator =(const generatorTemplate& a);
